Updated some validation
