+++ Midterm2020/Midterm2020/Status.cs
@@ -1,0 +1,8 @@
+﻿namespace Midterm2020
+{
+    public enum Status
+    {
+        OnShelf = 0,
+        CheckedOut = 1
+    }
+}
